Only render text surfaces once
First I rendered the title text, as it never changes, and then use that same surface over and over again until the title is turned off.

Then whenever it goes to display a `message`, check if a surface has already been rendered for that specific message. If not, make one. If so, use the surface we stored in `msg_cache`.

This reduces the font rendering bottleneck, and makes sure we only render text once, as needed.